requirements.txt
langchain==0.1.1
langchain-openai==0.0.3
python-dotenv==1.0.0